activate をhandlerに
